Alice in flower shop
While exploring in Wonderland, Alice stops by a flower shop. Just like everything else in Wonderland, this shop is unusual. The shop owner explains to Alice how to buy:

"In this shop, you will see many vases, each of which contains a number of flowers. For every coin you spend, you can remove all the flowers from a vase. After removing all the flowers from a vase, half (if it was even) or floor Half (if odd) the number of flowers planted will grow back in the pot, and you can use this pot again by paying the next coin."

Alice has a limited number of coins and wants to spend all her coins on buying flowers. Help him to pick as many flowers as possible from these strange pots.

There are N vases in the flower shop. In the first line of the input, N numbers are given, each of which is between 1 and 1 billion. In the second line, the number of Alice coins is specified.

In the only output line, you are expected to print the remainder of the maximum number of flowers that Alice can buy as 100000003. (Because the maximum number of flowers that Alice can buy may become so large that it exceeds the capacity of the variables!)

##input
66 84 84 94 3
7
##output
459

##input
74 90 22 34 2 82 61 11
7

##output
430